+++ common/src/main/java/org/exodusstudio/stellaris/common/blocks/entities/machines/GravityManipulatorBlockEntity.java
@@ -38,9 +38,8 @@ public class GravityManipulatorBlockEntity extends BaseEnergyContainerBlockEntit
 
     @Override
     public void tick(Level level, BlockState state) {
-        if  (isActive() && this.level != null) {
+        if (isActive()) {
             this.energyContainer.extract(Stellaris.CONFIG.gravityConfig.gravityManipulatorEnergyPerTick, false);
-            syncDataAccess();
         }
     }
 
@@ -85,14 +84,16 @@ public class GravityManipulatorBlockEntity extends BaseEnergyContainerBlockEntit
         return Mth.clamp((this.gravity - getMinGravity()) / range, 0.0, 1.0);
     }
 
-    public void setGravity(double gravity, boolean shouldSyncC2S) {
-        if (Math.abs(getGravity() - gravity) < 0.01) {
+    public void setGravity(double gravity, boolean shouldSync) {
+        double clamped = Mth.clamp(gravity, getMinGravity(), getMaxGravity());
+        if (Math.abs(getGravity() - clamped) < 0.01) {
             return;
         }
 
-        this.gravity = Mth.clamp(gravity, getMinGravity(), getMaxGravity());
+        this.gravity = clamped;
+        setChanged();
 
-        if  (shouldSyncC2S) {
+        if (shouldSync) {
             syncDataAccess();
         }
     }